+++ R/utils-yml.R
@@ -1,12 +1,10 @@
 # YAML Utilities
 #
-# Internal helpers for finding, reading, and writing YAML files.
+# Internal helpers for locating and validating YAML files.
 #
 # @param path `character(1)` Directory to search or write to.
 # @param name `character(1)` File name prefix (e.g. `"_mighty"`).
 # @param schema `character(1)` Path to the JSON schema file.
-# @param file `character(1)` or `NULL`. Path to a YAML file.
-# @param x Object to write (converted via `S7schema::to_yaml()`).
 
 #' @noRd
 list_yml <- function(path, name) {
@@ -38,14 +36,6 @@ find_yml <- function(path, name, schema) {
 }
 
 #' @noRd
-read_yml <- function(file) {
-  if (is.null(file)) {
-    return(list())
-  }
-  yaml::read_yaml(file = file)
-}
-
-#' @noRd
 match_yml <- function(path, name) {
   files <- list_yml(path, name)
 
@@ -54,15 +44,4 @@ match_yml <- function(path, name) {
   }
 
   files
-}
-
-#' @noRd
-write_yml <- function(x, path, name) {
-  cat(
-    S7schema::to_yaml(x),
-    file = match_yml(path, name),
-    sep = ""
-  )
-
-  invisible(x)
 }